server/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
requests>=2.28
python-dotenv>=1.0  # if using .env for configuration
pandas>=1.5  # if data processing is needed